requirements.txt
Flask==2.1.3
python-dotenv==0.21.0
psycopg2==2.9.5